pyproject.toml
[project]
name = "visionx"
version = "0.1.0"
description = ""
authors = [
    {name = "Gaurav Vijay Jadhav",email = "gaurav.vjadhav01@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "numpy (<2.3.0)",
    "albumentations (>=2.0.8,<3.0.0)",
    "faiss-cpu (>=1.11.0.post1,<2.0.0)",
    "pillow-heif (>=1.0.0,<2.0.0)",
    "rawpy (>=0.25.0,<0.26.0)",
    "transformers (>=4.53.3,<5.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "torchaudio (>=2.7.1,<3.0.0)",
    "onnxruntime-gpu (>=1.22.0,<2.0.0)",
]

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch-gpu"}
torchvision = {source = "pytorch-gpu"}
torchaudio = {source = "pytorch-gpu"}
[tool.poetry]
packages = [{include = "visionx", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
